+++ genium/src/test/resources/smoke/nullable/output/android/jni/StructConversion.cpp
@@ -255,5 +255,55 @@ convert_to_jni(JNIEnv* _jenv, const ::genium::optional<::smoke::Nullable::SomeSt
 {
     return _ninput ? convert_to_jni(_jenv, *_ninput) : JniReference<jobject>{};
 }
+::smoke::NullableCollectionsStruct
+convert_from_jni(JNIEnv* _jenv, const JniReference<jobject>& _jinput, ::smoke::NullableCollectionsStruct* dummy)
+{
+    ::smoke::NullableCollectionsStruct _nout{};
+    ::std::vector< ::genium::optional< ::std::chrono::system_clock::time_point > > n_dates = convert_from_jni(
+        _jenv,
+        ::genium::jni::get_object_field_value(
+        _jenv,
+        _jinput,
+        "dates",
+        "Ljava/util/List;"),
+        (::std::vector< ::genium::optional< ::std::chrono::system_clock::time_point > >*)nullptr );
+    _nout.dates = n_dates;
+    ::std::unordered_map< int32_t, ::genium::optional< ::smoke::Nullable::SomeStruct > > n_structs = convert_from_jni(
+        _jenv,
+        ::genium::jni::get_object_field_value(
+        _jenv,
+        _jinput,
+        "structs",
+        "Ljava/util/Map;"),
+        (::std::unordered_map< int32_t, ::genium::optional< ::smoke::Nullable::SomeStruct > >*)nullptr );
+    _nout.structs = n_structs;
+    return _nout;
+}
+::genium::optional<::smoke::NullableCollectionsStruct>
+convert_from_jni(JNIEnv* _jenv, const JniReference<jobject>& _jinput, ::genium::optional<::smoke::NullableCollectionsStruct>* dummy)
+{
+    return _jinput
+        ? ::genium::optional<::smoke::NullableCollectionsStruct>(convert_from_jni(_jenv, _jinput, (::smoke::NullableCollectionsStruct*)nullptr))
+        : ::genium::optional<::smoke::NullableCollectionsStruct>{};
+}
+REGISTER_JNI_CLASS_CACHE("com/example/smoke/NullableCollectionsStruct", ::smoke::NullableCollectionsStruct)
+JniReference<jobject>
+convert_to_jni(JNIEnv* _jenv, const ::smoke::NullableCollectionsStruct& _ninput)
+{
+    auto& javaClass = CachedJavaClass<::smoke::NullableCollectionsStruct>::java_class;
+    auto _jresult = ::genium::jni::alloc_object(_jenv, javaClass);
+    auto jdates = convert_to_jni(_jenv, _ninput.dates);
+    ::genium::jni::set_object_field_value(_jenv, _jresult, "dates",
+        "Ljava/util/List;", jdates);
+    auto jstructs = convert_to_jni(_jenv, _ninput.structs);
+    ::genium::jni::set_object_field_value(_jenv, _jresult, "structs",
+        "Ljava/util/Map;", jstructs);
+    return _jresult;
+}
+JniReference<jobject>
+convert_to_jni(JNIEnv* _jenv, const ::genium::optional<::smoke::NullableCollectionsStruct> _ninput)
+{
+    return _ninput ? convert_to_jni(_jenv, *_ninput) : JniReference<jobject>{};
+}
 }
 }